requirements.txt
Django==3.1.4
djangorestframework==3.12.2
requests==2.25.1
django-cors-headers==3.6.0
django-storages==1.11.1
python-dotenv==0.15.0
boto3==1.16.47
drf-yasg2==1.19.4
